Removed old skill stuff
