+++ client/src/components/characters/character/skills/SkillDetail.vue
@@ -1,70 +1,84 @@
 <script setup lang="ts">
-import {computed, type PropType} from 'vue';
-import SkeletonWrapper from "@/FormWrappers/SkeletonWrapper.vue";
-import type {SkillResponse} from "@/components/characters/character/skills/interfaces/SkillOptionsResponse";
-import {skillStore} from "@/components/characters/character/skills/Stores/skillStore";
 
-const skillInfo = skillStore();
+import {computed, onMounted, ref, type Ref} from "vue";
+import type {SkillResponse} from "@/components/characters/character/skills/interfaces/SkillOptionsResponse";
+import {useRoute} from 'vue-router'
+import {FeatureFlags, userStore} from "@/stores/userStore.ts";
+import axios from "axios";
+import SkeletonWrapper from "@/FormWrappers/SkeletonWrapper.vue";
+
+const route = useRoute();
+
 const props = defineProps({
-  selectedItem: {
-    type: Object as PropType<SkillResponse>,
+  skillTypeId: {
+    type: Number,
     required: true,
   },
-  isLoading: {
-    type: Boolean,
-    required: true
-  },
-  currentXpLevel:{
-    type: Number
-  },
-  showcaseOnly:{
-    type: Boolean
+  selectedLevelId:{
+    type: Number,
+    required: true,
   }
 });
 
-const plusOrMinusSign = computed(() => {
-  return props.selectedItem.experienceCost > props.currentXpLevel ? "-" : "+";
+const skillLevels:Ref<Array<SkillResponse>> = ref([]);
+const isLoading = ref(true);
+const selectedItem = ref(props.selectedLevelId);
+const showCharacterWizard = ref(false);
+const userInfo = userStore();
+
+
+const selectedLevel = computed(() => {
+  return skillLevels.value.find(x => x.levelId === selectedItem.value)
+})
+
+onMounted(async () =>{
+  getEditOptions();
+  showCharacterWizard.value = await userInfo.hasFeatureFlag(FeatureFlags.ShowCharacterWizard);
 });
+
+function getEditOptions() {
+  axios.get(`characters/${route.params.id}/skills/${props.skillTypeId}`)
+      .then((response) => {
+        skillLevels.value = response.data;
+        isLoading.value = false;
+      })
+}
 
 </script>
 
 <template>
-  <SkeletonWrapper :show-skeleton="props.isLoading" height="5rem" width="100%">
-    <div class="row">
-      <div class="col text-left">
-        <div class="mb-2">
-          Level {{ props.selectedItem.levelNumber }}
-        </div>
-        <div>
-          <SkeletonWrapper :show-skeleton="isLoading" height="2rem" width="100%">
-            {{ props.selectedItem.name }}
-          </SkeletonWrapper>
-        </div>
-      </div>
-      <div v-if="skillInfo.showExperience && !props.showcaseOnly" class="col text-right">
-        <div class="mb-2">
-          XP
-        </div>
-        <div>
-          <SkeletonWrapper :show-skeleton="isLoading" height="2rem" width="100%">
-            {{ plusOrMinusSign }} {{ Math.abs(props.selectedItem?.experienceCost - props.currentXpLevel) }}
-          </SkeletonWrapper>
-        </div>
-      </div>
-    </div>
-    <p class="m-0 mt-3 pb-2">
-      {{ props.selectedItem.description }}
-    </p>
-    <div v-if="props.selectedItem.benefits && props.selectedItem.benefits.length > 0">
-      <h3>Benefits</h3>
-      <div v-for="benefit in props.selectedItem.benefits">
-        <h4 class="d-flex justify-content-between w-100">
-          <div>{{ benefit.name }}</div>
-          <div class="text-right">
-            +{{ benefit.modifier }}
+  <div class="row pt-3">
+    <div class="col p-0 m-0">
+      <div class="p-3" >
+        <SkeletonWrapper :show-skeleton="isLoading" height="5rem" width="100%">
+          <div class="row">
+            <div class="col text-left">
+              <div class="mb-2">
+                Level {{ selectedLevel.levelNumber }}
+              </div>
+              <div>
+                <SkeletonWrapper :show-skeleton="isLoading" height="2rem" width="100%">
+                  {{ selectedLevel.name }}
+                </SkeletonWrapper>
+              </div>
+            </div>
           </div>
-        </h4>
+          <p class="m-0 mt-3 pb-2">
+            {{ selectedLevel.description }}
+          </p>
+          <div v-if="selectedLevel.benefits && selectedLevel.benefits.length > 0">
+            <h3>Benefits</h3>
+            <div v-for="benefit in selectedLevel.benefits">
+              <h4 class="d-flex justify-content-between w-100">
+                <div>{{ benefit.name }}</div>
+                <div class="text-right">
+                  +{{ benefit.modifier }}
+                </div>
+              </h4>
+            </div>
+          </div>
+        </SkeletonWrapper>
       </div>
     </div>
-  </SkeletonWrapper>
+  </div>
 </template>
